WIP: Unify short url for dashboards and explore
